+++ Elements/src/Search/TraversalConstraint.cs
@@ -1,0 +1,8 @@
+namespace Elements.Search
+{
+    internal enum TraversalConstraint
+    {
+        MinimumAngle,
+        MaximumAngle,
+    }
+}