+++ CHANGELOG.md
@@ -2,11 +2,9 @@
 
 ## 0.9.4
 - `LineSegmentExtensions.Intersections(this IList<Line> lines)`
-- `Elements.Search.BinaryTree<T>`
-- `Elements.Search.BinaryTreeNode<T>`
-- `Elements.Search.AdjacencyList<T>`
 - `Elements.Search.DistanceComparer`
 - `Elements.Search.DirectionComparer`
+- `Elements.Search.Network<T>`
 
 ## 0.9.3